add session 2.1

+++ mixnet/cmd/queuesim.py
@@ -11,7 +11,7 @@ if __name__ == "__main__":
     )
     parser.add_argument("--exp-id", type=int, required=True, help="Experiment ID (>=1)")
     parser.add_argument(
-        "--session-id", type=int, required=True, help="Session ID (>=1)"
+        "--session-id", type=float, required=True, help="Session ID (>=1)"
     )
     parser.add_argument(
         "--queue-type",
